Add files via upload
These are the docker build files however the SASHomeTar.tar absent. To build the SASHomeTar.tar file you will need to install SAS 9.4 on a similar build to the one referenced in the Docker file (I built it on centos7 (latest)) Then tart the SASHome folder and place it in your repository clone before building the docker image.